Move parsing package to correct location
