+++ components/inspectit-ocelot-config-doc-generator/src/main/java/inspectit/ocelot/config/doc/generator/ConfigDocManager.java
@@ -2,7 +2,7 @@ package inspectit.ocelot.config.doc.generator;
 
 import inspectit.ocelot.config.doc.generator.docobjects.ConfigDocumentation;
 import inspectit.ocelot.config.doc.generator.docobjects.DocObjectGenerator;
-import parsing.ConfigParser;
+import inspectit.ocelot.config.doc.generator.parsing.ConfigParser;
 import rocks.inspectit.ocelot.config.model.InspectitConfig;
 
 public class ConfigDocManager {
