+++ components/inspectit-ocelot-config-doc-generator/src/main/java/inspectit/ocelot/config/doc/generator/parsing/StringSubstitutorNestedMap.java
@@ -1,4 +1,4 @@
-package parsing;
+package inspectit.ocelot.config.doc.generator.parsing;
 
 import org.apache.commons.text.StringSubstitutor;
